Create Team Contribution
Group #24
Final Project - Team Contribution 

Working together as team to build our final program has been a great experience. We all were very organized and all willing to start working in advance as opposed to postponing things to the last minute. We really worked well as a group in terms of communication and being accepting of each others’ ideas. When it came to skills and knowledge of IST 256,we all were able to bring a sufficient amount of contribution to this project which was very helpful in the end. 

Steven was the highlight of our project after our brainstorming of ideas he was able to contribute a solid idea that we could actually be able to put into action. He was very organized and was able to bring to our first meeting a sufficient amount of code to give us a great headstart on our project. He had great communications skills and was always on time.

Joseph was also a great contribution to our assignment with his skills in photoshop he was able to design a great poster to represent our program. Within our end result of our program he contributed most to the beginning phases of bringing our map to the program. He also holds a great amount of technical skills overall and his insight was very helpful to the team. He was also very well with being timely and communication. 

Paige helped a lot with the organization and communication of our group - getting us to start focusing on our topics and setting up our meetings. When it came to the actual programming portion of the assignment, she holds a good amount of technical skills and was able to contribute external research for certain parts and in return Steven would put it into code as he was the beginner of the code. She has also contributed to this assignment by creating our demo video for the project.